Improved readability by moving the ref feature in a seperate file.
